Base BallotCard class
